requirements.txt
pytest
python-decouple
playwright~=1.40.0
allure-pytest
allure-combine

